# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FSP Talent API",
  "description": "\n**FSP Talent** — платформа подбора ИТ-специалистов с обратной механикой:\nработодатель сам находит кандидата в категоризированной базе и отправляет приглашение с вилкой ЗП.\n\n### Как авторизоваться в Swagger\n1. `POST /api/auth/register` → `POST /api/auth/verify-email` (в dev-режиме токен приходит в ответе регистрации).\n2. Кнопка **Authorize** вверху: username = email, password = пароль.\n\n### Формат ошибок\n`{\"code\": \"MACHINE_CODE\", \"message\": \"текст для пользователя\", \"detail\": \"...\"}`, для 422 ещё `fields: {поле: ошибка}`\n",
  "version": "1.3.0"
 },
 "paths": {
...
  "/api/testing/attempts": {
   "get": {
    "tags": [
     "Кандидат: тестирование"
    ],
    "summary": "История моих попыток",
    "operationId": "history_api_testing_attempts_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AttemptSummary"
         },
         "type": "array",
         "title": "Response History Api Testing Attempts Get"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Кандидат: тестирование"
    ],
    "summary": "Шаг 2–3: выбрать грейд и начать тест",
    "description": "Генерирует УНИКАЛЬНЫЙ вариант теста. Ошибки 409 (поле `code`):\nSURVEY_REQUIRED, ATTEMPT_IN_PROGRESS (+attempt_id), GRADE_COOLDOWN (+next_change_at), RETRY_COOLDOWN.",
    "operationId": "start_api_testing_attempts_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StartAttemptIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AttemptOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/testing/attempts/{attempt_id}": {
   "get": {
    "tags": [
     "Кандидат: тестирование"
    ],
    "summary": "Получить попытку (например, после перезагрузки страницы)",
    "operationId": "get_attempt_api_testing_attempts__attempt_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "attempt_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Attempt Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AttemptOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/testing/attempts/{attempt_id}/submit": {
   "post": {
    "tags": [
     "Кандидат: тестирование"
    ],
    "summary": "Отправить ответы и получить результат + решение по грейду",
    "operationId": "submit_api_testing_attempts__attempt_id__submit_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "attempt_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Attempt Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SubmitIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AttemptResultOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/testing/attempts/{attempt_id}/result": {
   "get": {
    "tags": [
     "Кандидат: тестирование"
    ],
    "summary": "Результат завершённой попытки",
    "operationId": "result_api_testing_attempts__attempt_id__result_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "attempt_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Attempt Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AttemptResultOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AttemptItemPublic": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id",
      "examples": [
       "q1"
      ]
     },
     "level": {
      "type": "integer",
      "title": "Level",
      "description": "0=Intern … 3=Senior"
     },
     "skill": {
      "type": "string",
      "title": "Skill"
     },
     "kind": {
      "type": "string",
      "title": "Kind",
      "description": "single — выбрать вариант, input — ввести ответ, code — написать функцию"
     },
     "text": {
      "type": "string",
      "title": "Text"
     },
     "options": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Options"
     },
     "code": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CodeSpec"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "level",
     "skill",
     "kind",
     "text"
    ],
    "title": "AttemptItemPublic",
    "description": "Задание БЕЗ правильного ответа."
   },
   "AttemptOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "specialization": {
      "type": "string",
      "title": "Specialization"
     },
     "target_grade": {
      "type": "string",
      "title": "Target Grade"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "started_at": {
      "type": "string",
      "format": "date-time",
      "title": "Started At"
     },
     "deadline_at": {
      "type": "string",
      "format": "date-time",
      "title": "Deadline At"
     },
     "seconds_left": {
      "type": "integer",
      "title": "Seconds Left"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/AttemptItemPublic"
      },
      "type": "array",
      "title": "Items"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score",
      "description": "Заполнено после сдачи"
     },
     "max_score": {
      "type": "integer",
      "title": "Max Score",
      "default": 100
     },
     "passed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Passed"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "specialization",
     "target_grade",
     "status",
     "started_at",
     "deadline_at",
     "seconds_left",
     "items"
    ],
    "title": "AttemptOut"
   },
   "AttemptResultOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "score": {
      "type": "number",
      "title": "Score"
     },
     "passed": {
      "type": "boolean",
      "title": "Passed"
     },
     "specialization": {
      "type": "string",
      "title": "Specialization"
     },
     "target_grade": {
      "type": "string",
      "title": "Target Grade"
     },
     "decision": {
      "$ref": "#/components/schemas/GradeDecision"
     },
     "by_skill": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "By Skill"
     },
     "by_level": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "By Level"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/ItemResult"
      },
      "type": "array",
      "title": "Items"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "status",
     "score",
     "passed",
     "specialization",
     "target_grade",
     "decision",
     "by_skill",
     "by_level",
     "items",
     "finished_at"
    ],
    "title": "AttemptResultOut"
   },
   "AttemptSummary": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "specialization": {
      "type": "string",
      "title": "Specialization"
     },
     "target_grade": {
      "type": "string",
      "title": "Target Grade"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     },
     "passed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Passed"
     },
     "started_at": {
      "type": "string",
      "format": "date-time",
      "title": "Started At"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "specialization",
     "target_grade",
     "status",
     "score",
     "passed",
     "started_at",
     "finished_at"
    ],
    "title": "AttemptSummary"
   },
...
   "CodeSpec": {
    "properties": {
     "function_name": {
      "type": "string",
      "title": "Function Name"
     },
     "signature": {
      "type": "string",
      "title": "Signature"
     },
     "examples": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Examples",
      "description": "Видимые примеры: [{args: [...], expected: ...}] — можно гонять в Pyodide"
     },
     "language": {
      "type": "string",
      "title": "Language",
      "default": "python"
     },
     "starter_code": {
      "type": "string",
      "title": "Starter Code",
      "description": "Что положить в редактор Monaco изначально"
     }
    },
    "type": "object",
    "required": [
     "function_name",
     "signature",
     "examples",
     "starter_code"
    ],
    "title": "CodeSpec"
   },
...
   "ErrorResponse": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "items": {},
        "type": "array"
       }
      ],
      "title": "Detail"
     },
     "fields": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "string"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fields"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "detail"
    ],
    "title": "ErrorResponse"
   },
...
   "GradeDecision": {
    "properties": {
     "grade_before": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grade Before"
     },
     "grade_after": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grade After"
     },
     "changed": {
      "type": "boolean",
      "title": "Changed"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "upgrade_suggested": {
      "type": "boolean",
      "title": "Upgrade Suggested",
      "default": false
     },
     "next_change_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Change At"
     }
    },
    "type": "object",
    "required": [
     "grade_before",
     "grade_after",
     "changed",
     "message"
    ],
    "title": "GradeDecision"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ItemResult": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "skill": {
      "type": "string",
      "title": "Skill"
     },
     "level": {
      "type": "integer",
      "title": "Level"
     },
     "score": {
      "type": "number",
      "title": "Score"
     },
     "status": {
      "type": "string",
      "title": "Status",
      "description": "correct | wrong | partial | no_answer"
     },
     "passed": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Passed"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     },
     "violations": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Violations"
     }
    },
    "type": "object",
    "required": [
     "id",
     "skill",
     "level",
     "score",
     "status"
    ],
    "title": "ItemResult"
   },
...
   "StartAttemptIn": {
    "properties": {
     "specialization": {
      "type": "string",
      "enum": [
       "backend",
       "frontend",
       "data_science",
       "qa",
       "devops"
      ],
      "title": "Specialization"
     },
     "target_grade": {
      "type": "string",
      "enum": [
       "intern",
       "junior",
       "middle",
       "senior"
      ],
      "title": "Target Grade",
      "description": "На какой грейд проходим тест"
     },
     "vacancy_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vacancy Id",
      "description": "Если тест строится под вакансию — задания подбираются по её стеку"
     }
    },
    "type": "object",
    "required": [
     "specialization",
     "target_grade"
    ],
    "title": "StartAttemptIn"
   },
...
   "SubmitIn": {
    "properties": {
     "answers": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Answers",
      "description": "id задания -> ответ. Для single — текст варианта, для code — исходный код",
      "examples": [
       {
        "q1": "201",
        "q2": "def solve(nums):\n    return 0"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "answers"
    ],
    "title": "SubmitIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/token"
     }
    }
   }
  }
 }
}